SSCS-8407 Adding ESA Scenario Integration Tests
