--- src/IntegrationTests/java/uk/gov/hmcts/reform/sscs/callback/WriteFinalDecisionIt.java
@@ -1,713 +1,0 @@
-package uk.gov.hmcts.reform.sscs.callback;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-import static uk.gov.hmcts.reform.sscs.ccd.callback.DocumentType.DRAFT_DECISION_NOTICE;
-import static uk.gov.hmcts.reform.sscs.helper.IntegrationTestHelper.assertHttpStatus;
-import static uk.gov.hmcts.reform.sscs.helper.IntegrationTestHelper.getRequestWithAuthHeader;
-
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.util.Collections;
-import org.junit.Assert;
-import org.junit.Test;
-import org.mockito.ArgumentCaptor;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.http.HttpStatus;
-import org.springframework.mock.web.MockHttpServletResponse;
-import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
-import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
-import uk.gov.hmcts.reform.idam.client.IdamClient;
-import uk.gov.hmcts.reform.idam.client.models.UserDetails;
-import uk.gov.hmcts.reform.sscs.ccd.callback.PreSubmitCallbackResponse;
-import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
-import uk.gov.hmcts.reform.sscs.docassembly.GenerateFile;
-import uk.gov.hmcts.reform.sscs.model.docassembly.GenerateFileParams;
-import uk.gov.hmcts.reform.sscs.model.docassembly.NoticeIssuedTemplateBody;
-import uk.gov.hmcts.reform.sscs.model.docassembly.WriteFinalDecisionTemplateBody;
-import uk.gov.hmcts.reform.sscs.service.EvidenceManagementService;
-
-@SpringBootTest
-@AutoConfigureMockMvc
-public class WriteFinalDecisionIt extends AbstractEventIt {
-
-    @MockBean
-    private CoreCaseDataApi coreCaseDataApi;
-
-    @MockBean
-    private IdamClient idamClient;
-
-    @MockBean
-    private AuthTokenGenerator authTokenGenerator;
-
-    @MockBean
-    private EvidenceManagementService evidenceManagementService;
-
-    @MockBean
-    private GenerateFile generateFile;
-
-    @MockBean
-    private UserDetails userDetails;
-
-    @Test
-    public void callToMidEventPreviewFinalDecisionCallback_willPreviewTheDocumentForDescriptorRoute() throws Exception {
-        setup();
-        setJsonAndReplace("callback/writeFinalDecisionDescriptor.json", "START_DATE_PLACEHOLDER", "2018-10-10");
-
-        String documentUrl = "document.url";
-        when(generateFile.assemble(any())).thenReturn(documentUrl);
-
-        when(userDetails.getFullName()).thenReturn("Judge Full Name");
-
-        when(idamClient.getUserDetails("Bearer userToken")).thenReturn(userDetails);
-
-        MockHttpServletResponse response = getResponse(getRequestWithAuthHeader(json, "/ccdMidEventPreviewFinalDecision"));
-        assertHttpStatus(response, HttpStatus.OK);
-        PreSubmitCallbackResponse<SscsCaseData> result = deserialize(response.getContentAsString());
-
-        assertEquals(Collections.EMPTY_SET, result.getErrors());
-
-        assertEquals(documentUrl, result.getData().getWriteFinalDecisionPreviewDocument().getDocumentUrl());
-
-        ArgumentCaptor<GenerateFileParams> capture = ArgumentCaptor.forClass(GenerateFileParams.class);
-        verify(generateFile).assemble(capture.capture());
-        final NoticeIssuedTemplateBody parentPayload = (NoticeIssuedTemplateBody) capture.getValue().getFormPayload();
-        final WriteFinalDecisionTemplateBody payload = parentPayload.getWriteFinalDecisionTemplateBody();
-
-        assertEquals("An Test", parentPayload.getAppellantFullName());
-        assertEquals("12345656789", parentPayload.getCaseId());
-        assertEquals("JT 12 34 56 D", parentPayload.getNino());
-        assertEquals("DRAFT DECISION NOTICE", parentPayload.getNoticeType());
-        assertEquals("Judge Full Name", parentPayload.getUserName());
-        assertEquals(LocalDate.parse("2017-07-17"), payload.getHeldOn());
-        assertEquals("Chester Magistrate's Court", payload.getHeldAt());
-        assertEquals("Judge Full Name, Panel Member 1 and Panel Member 2", payload.getHeldBefore());
-        assertEquals(true, payload.isAllowed());
-        assertEquals(true, payload.isSetAside());
-        assertEquals("2018-09-01", payload.getDateOfDecision());
-        assertEquals("An Test",payload.getAppellantName());
-        assertEquals("2018-10-10",payload.getStartDate());
-        assertEquals("2018-11-10",payload.getEndDate());
-        assertEquals(false, payload.isIndefinite());
-        assertEquals(true, payload.isDailyLivingIsEntited());
-        assertEquals(true, payload.isDailyLivingIsSeverelyLimited());
-        assertEquals("enhanced rate", payload.getDailyLivingAwardRate());
-        Assert.assertNotNull(payload.getDailyLivingDescriptors());
-        assertEquals(2, payload.getDailyLivingDescriptors().size());
-        Assert.assertNotNull(payload.getDailyLivingDescriptors().get(0));
-        assertEquals(8, payload.getDailyLivingDescriptors().get(0).getActivityAnswerPoints());
-        assertEquals("f", payload.getDailyLivingDescriptors().get(0).getActivityAnswerLetter());
-        assertEquals("Cannot prepare and cook food.", payload.getDailyLivingDescriptors().get(0).getActivityAnswerValue());
-        assertEquals("Preparing food", payload.getDailyLivingDescriptors().get(0).getActivityQuestionValue());
-        assertEquals("1", payload.getDailyLivingDescriptors().get(0).getActivityQuestionNumber());
-        Assert.assertNotNull(payload.getDailyLivingDescriptors().get(1));
-        assertEquals(10, payload.getDailyLivingDescriptors().get(1).getActivityAnswerPoints());
-        assertEquals("f", payload.getDailyLivingDescriptors().get(1).getActivityAnswerLetter());
-        assertEquals("Cannot convey food and drink to their mouth and needs another person to do so.", payload.getDailyLivingDescriptors().get(1).getActivityAnswerValue());
-        assertEquals("Taking nutrition", payload.getDailyLivingDescriptors().get(1).getActivityQuestionValue());
-        assertEquals("2", payload.getDailyLivingDescriptors().get(1).getActivityQuestionNumber());
-        Assert.assertNotNull(payload.getDailyLivingNumberOfPoints());
-        assertEquals(18, payload.getDailyLivingNumberOfPoints().intValue());
-        assertEquals(false, payload.isMobilityIsEntited());
-        assertEquals(false, payload.isMobilityIsSeverelyLimited());
-        Assert.assertNull(payload.getMobilityAwardRate());
-        Assert.assertNull(payload.getMobilityDescriptors());
-        assertNotNull(payload.getReasonsForDecision());
-        assertEquals(1, payload.getReasonsForDecision().size());
-        Assert.assertEquals("My reasons for decision", payload.getReasonsForDecision().get(0));
-        assertEquals("Something else.", payload.getAnythingElse());
-
-    }
-
-    @Test
-    public void callToMidEventPreviewFinalDecisionCallback_willPreviewTheDocumentForEsa() throws Exception {
-        setup();
-        setJsonAndReplace("callback/writeFinalDecisionDescriptorESA.json", "START_DATE_PLACEHOLDER", "2018-10-10");
-
-        String documentUrl = "document.url";
-        when(generateFile.assemble(any())).thenReturn(documentUrl);
-
-        when(userDetails.getFullName()).thenReturn("Judge Full Name");
-
-        when(idamClient.getUserDetails("Bearer userToken")).thenReturn(userDetails);
-
-        MockHttpServletResponse response = getResponse(getRequestWithAuthHeader(json, "/ccdMidEventPreviewFinalDecision"));
-        assertHttpStatus(response, HttpStatus.OK);
-        PreSubmitCallbackResponse<SscsCaseData> result = deserialize(response.getContentAsString());
-
-        assertEquals(Collections.EMPTY_SET, result.getErrors());
-
-        assertEquals(documentUrl, result.getData().getWriteFinalDecisionPreviewDocument().getDocumentUrl());
-
-        ArgumentCaptor<GenerateFileParams> capture = ArgumentCaptor.forClass(GenerateFileParams.class);
-        verify(generateFile).assemble(capture.capture());
-        final NoticeIssuedTemplateBody parentPayload = (NoticeIssuedTemplateBody) capture.getValue().getFormPayload();
-        final WriteFinalDecisionTemplateBody payload = parentPayload.getWriteFinalDecisionTemplateBody();
-
-        assertEquals("An Test", parentPayload.getAppellantFullName());
-        assertEquals("12345656789", parentPayload.getCaseId());
-        assertEquals("JT 12 34 56 D", parentPayload.getNino());
-        assertEquals("DRAFT DECISION NOTICE", parentPayload.getNoticeType());
-        assertEquals("Judge Full Name", parentPayload.getUserName());
-        assertEquals(LocalDate.parse("2017-07-17"), payload.getHeldOn());
-        assertEquals("Chester Magistrate's Court", payload.getHeldAt());
-        assertEquals("Judge Full Name, Panel Member 1 and Panel Member 2", payload.getHeldBefore());
-        assertEquals(true, payload.isAllowed());
-        assertEquals(true, payload.isSetAside());
-        assertEquals("2018-09-01", payload.getDateOfDecision());
-        assertEquals("An Test",payload.getAppellantName());
-        assertEquals("2018-10-10",payload.getStartDate());
-        assertEquals("2018-11-10",payload.getEndDate());
-        assertEquals(false, payload.isIndefinite());
-        assertEquals(true, payload.isEsaIsEntited());
-        assertEquals("lower rate", payload.getEsaAwardRate());
-        Assert.assertNotNull(payload.getEsaSchedule2Descriptors());
-        assertEquals(1, payload.getEsaSchedule2Descriptors().size());
-        assertEquals(15, payload.getEsaSchedule2Descriptors().get(0).getActivityAnswerPoints());
-        assertEquals("a", payload.getEsaSchedule2Descriptors().get(0).getActivityAnswerLetter());
-        assertEquals("Cannot, unaided by another person, either: (i) mobilise more than 50 metres on level ground without stopping in order to avoid significant discomfort or exhaustion; or (ii) repeatedly mobilise 50 metres within a reasonable timescale because of significant discomfort or exhaustion.", payload.getEsaSchedule2Descriptors().get(0).getActivityAnswerValue());
-        assertEquals("1. Mobilising unaided by another person with or without a walking stick, manual wheelchair or other aid if such aid is normally or could reasonably be worn or used.", payload.getEsaSchedule2Descriptors().get(0).getActivityQuestionValue());
-        assertEquals("1", payload.getEsaSchedule2Descriptors().get(0).getActivityQuestionNumber());
-        Assert.assertNotNull(payload.getEsaNumberOfPoints());
-        assertEquals(15, payload.getEsaNumberOfPoints().intValue());
-        assertNotNull(payload.getReasonsForDecision());
-        assertEquals(1, payload.getReasonsForDecision().size());
-        Assert.assertEquals("My reasons for decision", payload.getReasonsForDecision().get(0));
-        assertEquals("Something else.", payload.getAnythingElse());
-        assertNotNull(parentPayload.getWriteFinalDecisionTemplateContent());
-        assertNotNull(parentPayload.getWriteFinalDecisionTemplateContent());
-        assertEquals("The appeal is allowed.\n"
-            + "\n"
-            + "The decision made by the Secretary of State on 01/09/2018 is set aside.\n"
-            + "\n"
-            + "An Test has limited capability for work.\n"
-            + "\n"
-            + "In applying the Work Capability Assessment 15 points were scored from the activities and descriptors in Schedule 2 of the Employment and Support Allowance (ESA) Regulations 2008 made up as follows:\n"
-            + "\n"
-            + "1. Mobilising unaided by another person with or without a walking stick, manual wheelchair or other aid if such aid is normally or could reasonably be worn or used.\ta.Cannot, unaided by another person, either: (i) mobilise more than 50 metres on level ground without stopping in order to avoid significant discomfort or exhaustion; or (ii) repeatedly mobilise 50 metres within a reasonable timescale because of significant discomfort or exhaustion.\t15\n"
-            + "\n"
-            + "\n"
-            + "An Test does not have limited capability for work-related activity because no descriptor from Schedule 3 of the ESA Regulations applied. Regulation 35 did not apply.\n"
-            + "\n"
-            + "My reasons for decision\n"
-            + "\n"
-            + "Something else.\n"
-            + "\n"
-            + "This has been an oral (face to face) hearing. An Test attended the hearing today and the Tribunal considered the appeal bundle to page A1. A Presenting Officer attended on behalf of the Respondent.\n"
-            + "\n", parentPayload.getWriteFinalDecisionTemplateContent().toString());
-    }
-
-    @Test
-    public void callToMidEventPreviewFinalDecisionCallback_willPreviewTheDocumentForEsaNonWca_WhenPreviousQuestionsAnsweredAndOver15Points() throws Exception {
-        setup();
-        setJsonAndReplace("callback/writeFinalDecisionDescriptorESANonWCA.json", "START_DATE_PLACEHOLDER", "2018-10-10");
-
-        String documentUrl = "document.url";
-        when(generateFile.assemble(any())).thenReturn(documentUrl);
-
-        when(userDetails.getFullName()).thenReturn("Judge Full Name");
-
-        when(idamClient.getUserDetails("Bearer userToken")).thenReturn(userDetails);
-
-        MockHttpServletResponse response = getResponse(getRequestWithAuthHeader(json, "/ccdMidEventPreviewFinalDecision"));
-        assertHttpStatus(response, HttpStatus.OK);
-        PreSubmitCallbackResponse<SscsCaseData> result = deserialize(response.getContentAsString());
-
-        assertEquals(Collections.EMPTY_SET, result.getErrors());
-
-        assertEquals(documentUrl, result.getData().getWriteFinalDecisionPreviewDocument().getDocumentUrl());
-
-        ArgumentCaptor<GenerateFileParams> capture = ArgumentCaptor.forClass(GenerateFileParams.class);
-        verify(generateFile).assemble(capture.capture());
-        final NoticeIssuedTemplateBody parentPayload = (NoticeIssuedTemplateBody) capture.getValue().getFormPayload();
-        final WriteFinalDecisionTemplateBody payload = parentPayload.getWriteFinalDecisionTemplateBody();
-
-        assertEquals("An Test", parentPayload.getAppellantFullName());
-        assertEquals("12345656789", parentPayload.getCaseId());
-        assertEquals("JT 12 34 56 D", parentPayload.getNino());
-        assertEquals("DRAFT DECISION NOTICE", parentPayload.getNoticeType());
-        assertEquals("Judge Full Name", parentPayload.getUserName());
-        assertEquals(LocalDate.parse("2017-07-17"), payload.getHeldOn());
-        assertEquals("Chester Magistrate's Court", payload.getHeldAt());
-        assertEquals("Judge Full Name, Panel Member 1 and Panel Member 2", payload.getHeldBefore());
-        assertEquals(true, payload.isAllowed());
-        assertEquals(true, payload.isSetAside());
-        assertEquals("2018-09-01", payload.getDateOfDecision());
-        assertEquals("An Test",payload.getAppellantName());
-        assertEquals("2018-10-10",payload.getStartDate());
-        assertEquals("2018-11-10",payload.getEndDate());
-        assertEquals(false, payload.isIndefinite());
-        assertEquals(false, payload.isEsaIsEntited());
-        assertNull(payload.getEsaAwardRate());
-        Assert.assertNull(payload.getEsaSchedule2Descriptors());
-        Assert.assertNull(payload.getEsaNumberOfPoints());
-        assertNotNull(payload.getReasonsForDecision());
-        assertEquals(1, payload.getReasonsForDecision().size());
-        Assert.assertEquals("My reasons for decision", payload.getReasonsForDecision().get(0));
-        assertEquals("Something else.", payload.getAnythingElse());
-        assertNotNull(parentPayload.getWriteFinalDecisionTemplateContent());
-        assertNotNull(parentPayload.getWriteFinalDecisionTemplateContent());
-        assertEquals("The appeal is allowed.\n"
-            + "\n"
-            + "The decision made by the Secretary of State on 01/09/2018 is set aside.\n"
-            + "\n"
-            + "This is my summary.\n"
-            + "\n"
-            + "My reasons for decision\n"
-            + "\n"
-            + "Something else.\n"
-            + "\n"
-            + "This has been an oral (face to face) hearing. An Test attended the hearing today and the Tribunal considered the appeal bundle to page A1. A Presenting Officer attended on behalf of the Respondent.\n"
-            + "\n", parentPayload.getWriteFinalDecisionTemplateContent().toString());
-    }
-
-    @Test
-    public void callToMidEventPreviewFinalDecisionCallback_willPreviewTheDocumentForUc() throws Exception {
-        setup();
-        setJsonAndReplace("callback/writeFinalDecisionDescriptorUC.json", "START_DATE_PLACEHOLDER", "2018-10-10");
-
-        String documentUrl = "document.url";
-        when(generateFile.assemble(any())).thenReturn(documentUrl);
-
-        when(userDetails.getFullName()).thenReturn("Judge Full Name");
-
-        when(idamClient.getUserDetails("Bearer userToken")).thenReturn(userDetails);
-
-        MockHttpServletResponse response = getResponse(getRequestWithAuthHeader(json, "/ccdMidEventPreviewFinalDecision"));
-        assertHttpStatus(response, HttpStatus.OK);
-        PreSubmitCallbackResponse<SscsCaseData> result = deserialize(response.getContentAsString());
-
-        assertEquals(Collections.EMPTY_SET, result.getErrors());
-
-        assertEquals(documentUrl, result.getData().getWriteFinalDecisionPreviewDocument().getDocumentUrl());
-
-        ArgumentCaptor<GenerateFileParams> capture = ArgumentCaptor.forClass(GenerateFileParams.class);
-        verify(generateFile).assemble(capture.capture());
-        final NoticeIssuedTemplateBody parentPayload = (NoticeIssuedTemplateBody) capture.getValue().getFormPayload();
-        final WriteFinalDecisionTemplateBody payload = parentPayload.getWriteFinalDecisionTemplateBody();
-
-        assertEquals("An Test", parentPayload.getAppellantFullName());
-        assertEquals("12345656789", parentPayload.getCaseId());
-        assertEquals("JT 12 34 56 D", parentPayload.getNino());
-        assertEquals("DRAFT DECISION NOTICE", parentPayload.getNoticeType());
-        assertEquals("Judge Full Name", parentPayload.getUserName());
-        assertEquals(LocalDate.parse("2017-07-17"), payload.getHeldOn());
-        assertEquals("Chester Magistrate's Court", payload.getHeldAt());
-        assertEquals("Judge Full Name, Panel Member 1 and Panel Member 2", payload.getHeldBefore());
-        assertEquals(true, payload.isAllowed());
-        assertEquals(true, payload.isSetAside());
-        assertEquals("2018-09-01", payload.getDateOfDecision());
-        assertEquals("An Test",payload.getAppellantName());
-        assertEquals("2018-10-10",payload.getStartDate());
-        assertEquals("2018-11-10",payload.getEndDate());
-        assertEquals(false, payload.isIndefinite());
-        assertEquals(true, payload.isUcIsEntited());
-        assertEquals("lower rate", payload.getUcAwardRate());
-        Assert.assertNotNull(payload.getUcSchedule6Descriptors());
-        assertEquals(1, payload.getUcSchedule6Descriptors().size());
-        assertEquals(15, payload.getUcSchedule6Descriptors().get(0).getActivityAnswerPoints());
-        assertEquals("a", payload.getUcSchedule6Descriptors().get(0).getActivityAnswerLetter());
-        assertEquals("Cannot, unaided by another person, either: (i) mobilise more than 50 metres on level ground without stopping in order to avoid significant discomfort or exhaustion; or (ii) repeatedly mobilise 50 metres within a reasonable timescale because of significant discomfort or exhaustion.", payload.getUcSchedule6Descriptors().get(0).getActivityAnswerValue());
-        assertEquals("1. Mobilising unaided by another person with or without a walking stick, manual wheelchair or other aid if such aid is normally or could reasonably be worn or used.", payload.getUcSchedule6Descriptors().get(0).getActivityQuestionValue());
-        assertEquals("1", payload.getUcSchedule6Descriptors().get(0).getActivityQuestionNumber());
-        Assert.assertNotNull(payload.getUcNumberOfPoints());
-        assertEquals(15, payload.getUcNumberOfPoints().intValue());
-        assertNotNull(payload.getReasonsForDecision());
-        assertEquals(1, payload.getReasonsForDecision().size());
-        Assert.assertEquals("My reasons for decision", payload.getReasonsForDecision().get(0));
-        assertEquals("Something else.", payload.getAnythingElse());
-        assertNotNull(parentPayload.getWriteFinalDecisionTemplateContent());
-        assertEquals("The appeal is allowed.\n"
-            + "\n"
-            + "The decision made by the Secretary of State on 01/09/2018 is set aside.\n"
-            + "\n"
-            + "An Test has limited capability for work. The matter is now remitted to the Secretary of State to make a final decision upon entitlement to Universal Credit (UC).\n"
-            + "\n"
-            + "In applying the Work Capability Assessment 15 points were scored from the activities and descriptors in Schedule 6 of the UC Regulations 2013 made up as follows:\n"
-            + "\n"
-            + "1. Mobilising unaided by another person with or without a walking stick, manual wheelchair or other aid if such aid is normally or could reasonably be worn or used.\ta.Cannot, unaided by another person, either: (i) mobilise more than 50 metres on level ground without stopping in order to avoid significant discomfort or exhaustion; or (ii) repeatedly mobilise 50 metres within a reasonable timescale because of significant discomfort or exhaustion.\t15\n"
-            + "\n"
-            + "\n"
-            + "An Test does not have limited capability for work-related activity because no descriptor from Schedule 7 of the UC Regulations applied. Schedule 9, paragraph 4 did not apply.\n"
-            + "\n"
-            + "My reasons for decision\n"
-            + "\n"
-            + "Something else.\n"
-            + "\n"
-            + "This has been an oral (face to face) hearing. An Test attended the hearing today and the Tribunal considered the appeal bundle to page A1. A Presenting Officer attended on behalf of the Respondent.\n"
-            + "\n", parentPayload.getWriteFinalDecisionTemplateContent().toString());
-    }
-
-    @Test
-    public void callToMidEventPreviewFinalDecisionCallback_willPreviewTheDocumentForUc_WhenNonLcwa_WhenQuestionsPreviouslyAnsweredAndOver15Points() throws Exception {
-        setup();
-        setJsonAndReplace("callback/writeFinalDecisionDescriptorUCNonLcwa.json", "START_DATE_PLACEHOLDER", "2018-10-10");
-
-        String documentUrl = "document.url";
-        when(generateFile.assemble(any())).thenReturn(documentUrl);
-
-        when(userDetails.getFullName()).thenReturn("Judge Full Name");
-
-        when(idamClient.getUserDetails("Bearer userToken")).thenReturn(userDetails);
-
-        MockHttpServletResponse response = getResponse(getRequestWithAuthHeader(json, "/ccdMidEventPreviewFinalDecision"));
-        assertHttpStatus(response, HttpStatus.OK);
-        PreSubmitCallbackResponse<SscsCaseData> result = deserialize(response.getContentAsString());
-
-        assertEquals(Collections.EMPTY_SET, result.getErrors());
-
-        assertEquals(documentUrl, result.getData().getWriteFinalDecisionPreviewDocument().getDocumentUrl());
-
-        ArgumentCaptor<GenerateFileParams> capture = ArgumentCaptor.forClass(GenerateFileParams.class);
-        verify(generateFile).assemble(capture.capture());
-        final NoticeIssuedTemplateBody parentPayload = (NoticeIssuedTemplateBody) capture.getValue().getFormPayload();
-        final WriteFinalDecisionTemplateBody payload = parentPayload.getWriteFinalDecisionTemplateBody();
-
-        assertEquals("An Test", parentPayload.getAppellantFullName());
-        assertEquals("12345656789", parentPayload.getCaseId());
-        assertEquals("JT 12 34 56 D", parentPayload.getNino());
-        assertEquals("DRAFT DECISION NOTICE", parentPayload.getNoticeType());
-        assertEquals("Judge Full Name", parentPayload.getUserName());
-        assertEquals(LocalDate.parse("2017-07-17"), payload.getHeldOn());
-        assertEquals("Chester Magistrate's Court", payload.getHeldAt());
-        assertEquals("Judge Full Name, Panel Member 1 and Panel Member 2", payload.getHeldBefore());
-        assertEquals(true, payload.isAllowed());
-        assertEquals(true, payload.isSetAside());
-        assertEquals("2018-09-01", payload.getDateOfDecision());
-        assertEquals("An Test",payload.getAppellantName());
-        assertEquals("2018-10-10",payload.getStartDate());
-        assertEquals("2018-11-10",payload.getEndDate());
-        assertEquals(false, payload.isIndefinite());
-        assertEquals(false, payload.isUcIsEntited());
-        assertNull(payload.getUcAwardRate());
-        Assert.assertNull(payload.getUcSchedule6Descriptors());
-        Assert.assertNull(payload.getUcNumberOfPoints());
-        assertNotNull(payload.getReasonsForDecision());
-        assertEquals(1, payload.getReasonsForDecision().size());
-        Assert.assertEquals("My reasons for decision", payload.getReasonsForDecision().get(0));
-        assertEquals("Something else.", payload.getAnythingElse());
-        assertNotNull(parentPayload.getWriteFinalDecisionTemplateContent());
-        assertEquals("The appeal is allowed.\n"
-            + "\n"
-            + "The decision made by the Secretary of State on 01/09/2018 is set aside.\n"
-            + "\n"
-            + "This is my summary.\n"
-            + "\n"
-            + "My reasons for decision\n"
-            + "\n"
-            + "Something else.\n"
-            + "\n"
-            + "This has been an oral (face to face) hearing. An Test attended the hearing today and the Tribunal considered the appeal bundle to page A1. A Presenting Officer attended on behalf of the Respondent.\n"
-            + "\n", parentPayload.getWriteFinalDecisionTemplateContent().toString());
-    }
-
-    @Test
-    public void callToMidEventPreviewFinalDecisionCallback_willPreviewTheDocumentForNotConsideredDescriptorRoute() throws Exception {
-        setup();
-        setJsonAndReplace("callback/writeFinalDecisionDescriptorNotConsidered.json", "START_DATE_PLACEHOLDER", "2018-10-10");
-
-        String documentUrl = "document.url";
-        when(generateFile.assemble(any())).thenReturn(documentUrl);
-
-        when(userDetails.getFullName()).thenReturn("Judge Full Name");
-
-        when(idamClient.getUserDetails("Bearer userToken")).thenReturn(userDetails);
-
-        MockHttpServletResponse response = getResponse(getRequestWithAuthHeader(json, "/ccdMidEventPreviewFinalDecision"));
-        assertHttpStatus(response, HttpStatus.OK);
-        PreSubmitCallbackResponse<SscsCaseData> result = deserialize(response.getContentAsString());
-
-        assertEquals(Collections.EMPTY_SET, result.getErrors());
-
-        assertEquals(documentUrl, result.getData().getWriteFinalDecisionPreviewDocument().getDocumentUrl());
-
-        ArgumentCaptor<GenerateFileParams> capture = ArgumentCaptor.forClass(GenerateFileParams.class);
-        verify(generateFile).assemble(capture.capture());
-        final NoticeIssuedTemplateBody parentPayload = (NoticeIssuedTemplateBody) capture.getValue().getFormPayload();
-        final WriteFinalDecisionTemplateBody payload = parentPayload.getWriteFinalDecisionTemplateBody();
-
-        assertEquals("An Test", parentPayload.getAppellantFullName());
-        assertEquals("12345656789", parentPayload.getCaseId());
-        assertEquals("JT 12 34 56 D", parentPayload.getNino());
-        assertEquals("DRAFT DECISION NOTICE", parentPayload.getNoticeType());
-        assertEquals("Judge Full Name", parentPayload.getUserName());
-        assertEquals(LocalDate.parse("2017-07-17"), payload.getHeldOn());
-        assertEquals("Chester Magistrate's Court", payload.getHeldAt());
-        assertEquals("Judge Full Name, Panel Member 1 and Panel Member 2", payload.getHeldBefore());
-        assertEquals(true, payload.isAllowed());
-        assertEquals(true, payload.isSetAside());
-        assertEquals("2018-09-01", payload.getDateOfDecision());
-        assertEquals("An Test",payload.getAppellantName());
-        assertEquals("2018-10-10",payload.getStartDate());
-        assertEquals("2018-11-10",payload.getEndDate());
-        assertEquals(false, payload.isIndefinite());
-        assertEquals(true, payload.isDailyLivingIsEntited());
-        assertEquals(true, payload.isDailyLivingIsSeverelyLimited());
-        assertEquals("enhanced rate", payload.getDailyLivingAwardRate());
-        Assert.assertNotNull(payload.getDailyLivingDescriptors());
-        assertEquals(2, payload.getDailyLivingDescriptors().size());
-        Assert.assertNotNull(payload.getDailyLivingDescriptors().get(0));
-        assertEquals(8, payload.getDailyLivingDescriptors().get(0).getActivityAnswerPoints());
-        assertEquals("f", payload.getDailyLivingDescriptors().get(0).getActivityAnswerLetter());
-        assertEquals("Cannot prepare and cook food.", payload.getDailyLivingDescriptors().get(0).getActivityAnswerValue());
-        assertEquals("Preparing food", payload.getDailyLivingDescriptors().get(0).getActivityQuestionValue());
-        assertEquals("1", payload.getDailyLivingDescriptors().get(0).getActivityQuestionNumber());
-        Assert.assertNotNull(payload.getDailyLivingDescriptors().get(1));
-        assertEquals(10, payload.getDailyLivingDescriptors().get(1).getActivityAnswerPoints());
-        assertEquals("f", payload.getDailyLivingDescriptors().get(1).getActivityAnswerLetter());
-        assertEquals("Cannot convey food and drink to their mouth and needs another person to do so.", payload.getDailyLivingDescriptors().get(1).getActivityAnswerValue());
-        assertEquals("Taking nutrition", payload.getDailyLivingDescriptors().get(1).getActivityQuestionValue());
-        assertEquals("2", payload.getDailyLivingDescriptors().get(1).getActivityQuestionNumber());
-        Assert.assertNotNull(payload.getDailyLivingNumberOfPoints());
-        assertEquals(18, payload.getDailyLivingNumberOfPoints().intValue());
-        assertEquals(false, payload.isMobilityIsEntited());
-        assertEquals(false, payload.isMobilityIsSeverelyLimited());
-        Assert.assertNotNull(payload.getMobilityAwardRate());
-        assertEquals("not considered", payload.getMobilityAwardRate());
-        Assert.assertNull(payload.getMobilityDescriptors());
-        assertNotNull(payload.getReasonsForDecision());
-        assertEquals(1, payload.getReasonsForDecision().size());
-        Assert.assertEquals("My reasons for decision", payload.getReasonsForDecision().get(0));
-        assertEquals("Something else.", payload.getAnythingElse());
-
-    }
-
-    @Test
-    public void callToMidEventPreviewFinalDecisionCallback_willPreviewTheDocumentForNonDescriptorRoute() throws Exception {
-        setup();
-        setJsonAndReplace("callback/writeFinalDecisionNonDescriptor.json", "START_DATE_PLACEHOLDER", "2018-10-10");
-
-        String documentUrl = "document.url";
-        when(generateFile.assemble(any())).thenReturn(documentUrl);
-
-        when(userDetails.getFullName()).thenReturn("Judge Full Name");
-
-        when(idamClient.getUserDetails("Bearer userToken")).thenReturn(userDetails);
-
-        MockHttpServletResponse response = getResponse(getRequestWithAuthHeader(json, "/ccdMidEventPreviewFinalDecision"));
-        assertHttpStatus(response, HttpStatus.OK);
-        PreSubmitCallbackResponse<SscsCaseData> result = deserialize(response.getContentAsString());
-
-        assertEquals(Collections.EMPTY_SET, result.getErrors());
-
-        assertEquals(documentUrl, result.getData().getWriteFinalDecisionPreviewDocument().getDocumentUrl());
-
-        ArgumentCaptor<GenerateFileParams> capture = ArgumentCaptor.forClass(GenerateFileParams.class);
-        verify(generateFile).assemble(capture.capture());
-        final NoticeIssuedTemplateBody parentPayload = (NoticeIssuedTemplateBody) capture.getValue().getFormPayload();
-        final WriteFinalDecisionTemplateBody payload = parentPayload.getWriteFinalDecisionTemplateBody();
-
-        assertEquals("An Test", parentPayload.getAppellantFullName());
-        assertEquals("12345656789", parentPayload.getCaseId());
-        assertEquals("JT 12 34 56 D", parentPayload.getNino());
-        assertEquals("DRAFT DECISION NOTICE", parentPayload.getNoticeType());
-        assertEquals("Judge Full Name", parentPayload.getUserName());
-        assertEquals(LocalDate.parse("2017-07-17"), payload.getHeldOn());
-        assertEquals("Chester Magistrate's Court", payload.getHeldAt());
-        assertEquals("Judge Full Name, Panel Member 1 and Panel Member 2", payload.getHeldBefore());
-        assertEquals(true, payload.isAllowed());
-        assertEquals(true, payload.isSetAside());
-        assertEquals("2018-09-01", payload.getDateOfDecision());
-        assertEquals("An Test",payload.getAppellantName());
-        assertNull(payload.getStartDate());
-        assertNull(payload.getEndDate());
-        assertEquals(true, payload.isIndefinite());
-        assertEquals(false, payload.isDailyLivingIsEntited());
-        assertEquals(false, payload.isDailyLivingIsSeverelyLimited());
-        assertNull(payload.getDailyLivingAwardRate());
-        assertNull(payload.getDailyLivingDescriptors());
-        Assert.assertNull(payload.getDailyLivingNumberOfPoints());
-        assertEquals(false, payload.isMobilityIsEntited());
-        assertEquals(false, payload.isMobilityIsSeverelyLimited());
-        assertNull(payload.getMobilityAwardRate());
-        assertNull(payload.getMobilityDescriptors());
-        assertNotNull(payload.getReasonsForDecision());
-        assertEquals(1, payload.getReasonsForDecision().size());
-        Assert.assertEquals("My reasons for decision", payload.getReasonsForDecision().get(0));
-        assertNotNull(payload.getDetailsOfDecision());
-        Assert.assertEquals("The details of the decision", payload.getDetailsOfDecision());
-        assertEquals("Something else.", payload.getAnythingElse());
-    }
-
-    @Test
-    public void callToMidEventCallback_willValidateTheDatePip() throws Exception {
-        setup();
-        setJsonAndReplace("callback/writeFinalDecisionDescriptor.json", "START_DATE_PLACEHOLDER", "2019-10-10");
-
-        MockHttpServletResponse response = getResponse(getRequestWithAuthHeader(json, "/ccdMidEvent"));
-        assertHttpStatus(response, HttpStatus.OK);
-        PreSubmitCallbackResponse<SscsCaseData> result = deserialize(response.getContentAsString());
-
-        assertEquals(1, result.getErrors().size());
-        assertEquals("Decision notice end date must be after decision notice start date", result.getErrors().toArray()[0]);
-
-    }
-
-    @Test
-    public void callToMidEventCallback_willValidateTheDateEsa() throws Exception {
-        setup();
-        setJsonAndReplace("callback/writeFinalDecisionDescriptorESA.json", "START_DATE_PLACEHOLDER", "2019-10-10");
-
-        MockHttpServletResponse response = getResponse(getRequestWithAuthHeader(json, "/ccdMidEvent"));
-        assertHttpStatus(response, HttpStatus.OK);
-        PreSubmitCallbackResponse<SscsCaseData> result = deserialize(response.getContentAsString());
-
-        assertEquals(1, result.getErrors().size());
-        assertEquals("Decision notice end date must be after decision notice start date", result.getErrors().toArray()[0]);
-
-    }
-
-    @Test
-    public void callToMidEventCallback_willValidateTheDateUc() throws Exception {
-        setup();
-        setJsonAndReplace("callback/writeFinalDecisionDescriptorUC.json", "START_DATE_PLACEHOLDER", "2019-10-10");
-
-        MockHttpServletResponse response = getResponse(getRequestWithAuthHeader(json, "/ccdMidEvent"));
-        assertHttpStatus(response, HttpStatus.OK);
-        PreSubmitCallbackResponse<SscsCaseData> result = deserialize(response.getContentAsString());
-
-        assertEquals(1, result.getErrors().size());
-        assertEquals("Decision notice end date must be after decision notice start date", result.getErrors().toArray()[0]);
-
-    }
-
-    @Test
-    public void callToAboutToSubmitHandler_willWriteDraftFinalDecisionToCase() throws Exception {
-        setup();
-        setJsonAndReplace("callback/writeFinalDecisionDescriptor.json", "START_DATE_PLACEHOLDER", "2018-10-10");
-
-        MockHttpServletResponse response = getResponse(getRequestWithAuthHeader(json, "/ccdAboutToSubmit"));
-        assertHttpStatus(response, HttpStatus.OK);
-        PreSubmitCallbackResponse<SscsCaseData> result = deserialize(response.getContentAsString());
-
-        assertEquals(Collections.EMPTY_SET, result.getErrors());
-
-        assertNull(result.getData().getOutcome());
-
-        assertEquals(DRAFT_DECISION_NOTICE.getValue(), result.getData().getSscsDocument().get(0).getValue().getDocumentType());
-        assertEquals(LocalDate.now().toString(), result.getData().getSscsDocument().get(0).getValue().getDocumentDateAdded());
-        assertEquals("Draft Decision Notice generated on " + LocalDate.now().format(DateTimeFormatter.ofPattern("dd-MM-YYYY")) + ".pdf", result.getData().getSscsDocument().get(0).getValue().getDocumentFileName());
-    }
-
-    /**
-     * Due to a CCD bug ( https://tools.hmcts.net/jira/browse/RDM-8200 ) we have had
-     * to implement a workaround in WriteFinalDecisionAboutToSubmitHandler to set
-     * the generated date to now, even though it is already being determined by the
-     * preview document handler.  This is because on submission, the correct generated date
-     * (the one referenced in the preview document) is being overwritten to a null value.
-     * Once RDM-8200 is fixed and we remove the workaround, this test should be changed
-     * to assert that a "something has gone wrong" error is displayed instead of
-     * previewing the document, as a null generated date would indicate that the
-     * date in the preview document hasn't been set.
-     *
-     */
-    @Test
-    public void callToAboutToSubmitHandler_willWriteDraftFinalDecisionToCaseWithGeneratedDateAsNowWhenSubmittedGeneratedDateIsNull() throws Exception {
-        setup();
-        setJsonAndReplace("callback/writeFinalDecisionDescriptorNullGeneratedDate.json", "START_DATE_PLACEHOLDER", "2018-10-10");
-
-        MockHttpServletResponse response = getResponse(getRequestWithAuthHeader(json, "/ccdAboutToSubmit"));
-        assertHttpStatus(response, HttpStatus.OK);
-        PreSubmitCallbackResponse<SscsCaseData> result = deserialize(response.getContentAsString());
-
-        assertEquals(Collections.EMPTY_SET, result.getErrors());
-
-        assertNull(result.getData().getOutcome());
-
-
-        assertNotNull(result.getData().getWriteFinalDecisionGeneratedDate());
-        assertEquals(LocalDate.now().toString(), result.getData().getWriteFinalDecisionGeneratedDate());
-
-        assertEquals(DRAFT_DECISION_NOTICE.getValue(), result.getData().getSscsDocument().get(0).getValue().getDocumentType());
-        assertEquals(LocalDate.now().toString(), result.getData().getSscsDocument().get(0).getValue().getDocumentDateAdded());
-        assertEquals("Draft Decision Notice generated on " + LocalDate.now().format(DateTimeFormatter.ofPattern("dd-MM-YYYY")) + ".pdf", result.getData().getSscsDocument().get(0).getValue().getDocumentFileName());
-    }
-
-    /**
-     * This test asserts that whatever the value of the existing generated date from CCD
-     * submitted as part of the payload to the WriterFinalSubmissionAboutToSubmitHandler,
-     * then that date is updated to now() after the WriterFinalSubmissionAboutToSubmitHandler is called.
-     * This is due to a workaround we have implemented in the WriterFinalSubmissionAboutToSubmitHandler
-     *
-     */
-    @Test
-    public void callToAboutToSubmitHandler_willWriteDraftFinalDecisionToCaseWithGeneratedDateAsNowWhenSubmittedGeneratedDateIsSet() throws Exception {
-        setup();
-        setJsonAndReplace("callback/writeFinalDecisionDescriptorSetGeneratedDate.json", "START_DATE_PLACEHOLDER", "2018-10-10");
-
-        MockHttpServletResponse response = getResponse(getRequestWithAuthHeader(json, "/ccdAboutToSubmit"));
-        assertHttpStatus(response, HttpStatus.OK);
-        PreSubmitCallbackResponse<SscsCaseData> result = deserialize(response.getContentAsString());
-
-        assertEquals(Collections.EMPTY_SET, result.getErrors());
-
-        assertNull(result.getData().getOutcome());
-
-
-        assertNotNull(result.getData().getWriteFinalDecisionGeneratedDate());
-
-        assertEquals(LocalDate.now().toString(), result.getData().getWriteFinalDecisionGeneratedDate());
-
-        assertEquals(DRAFT_DECISION_NOTICE.getValue(), result.getData().getSscsDocument().get(0).getValue().getDocumentType());
-        assertEquals(LocalDate.now().toString(), result.getData().getSscsDocument().get(0).getValue().getDocumentDateAdded());
-        assertEquals("Draft Decision Notice generated on " + LocalDate.now().format(DateTimeFormatter.ofPattern("dd-MM-YYYY")) + ".pdf", result.getData().getSscsDocument().get(0).getValue().getDocumentFileName());
-    }
-
-    @Test
-    public void callToAboutToSubmitHandler_willWriteDraftFinalDecisionToCaseForNonDescriptorRoute() throws Exception {
-        setup();
-        setJsonAndReplace("callback/writeFinalDecisionNonDescriptor.json", "START_DATE_PLACEHOLDER", "2018-10-10");
-
-        MockHttpServletResponse response = getResponse(getRequestWithAuthHeader(json, "/ccdAboutToSubmit"));
-        assertHttpStatus(response, HttpStatus.OK);
-        PreSubmitCallbackResponse<SscsCaseData> result = deserialize(response.getContentAsString());
-
-        assertEquals(Collections.EMPTY_SET, result.getErrors());
-
-        assertNull(result.getData().getOutcome());
-
-        assertEquals(DRAFT_DECISION_NOTICE.getValue(), result.getData().getSscsDocument().get(0).getValue().getDocumentType());
-        assertEquals(LocalDate.now().toString(), result.getData().getSscsDocument().get(0).getValue().getDocumentDateAdded());
-        assertEquals("Draft Decision Notice generated on " + LocalDate.now().format(DateTimeFormatter.ofPattern("dd-MM-YYYY")) + ".pdf", result.getData().getSscsDocument().get(0).getValue().getDocumentFileName());
-    }
-
-    @Test
-    public void callToAboutToSubmitHandler_willWriteManuallyUploadedFinalDecisionToCase() throws Exception {
-        setup("callback/writeFinalDecisionManualUploadDescriptor.json");
-
-        MockHttpServletResponse response = getResponse(getRequestWithAuthHeader(json, "/ccdAboutToSubmit"));
-        assertHttpStatus(response, HttpStatus.OK);
-        PreSubmitCallbackResponse<SscsCaseData> result = deserialize(response.getContentAsString());
-
-        assertEquals(Collections.EMPTY_SET, result.getErrors());
-
-        assertNull(result.getData().getOutcome());
-
-        assertEquals(DRAFT_DECISION_NOTICE.getValue(), result.getData().getSscsDocument().get(0).getValue().getDocumentType());
-        assertEquals(LocalDate.now().toString(), result.getData().getSscsDocument().get(0).getValue().getDocumentDateAdded());
-        assertEquals("Draft Decision Notice generated on " + LocalDate.now().format(DateTimeFormatter.ofPattern("dd-MM-yyyy")) + ".pdf", result.getData().getSscsDocument().get(0).getValue().getDocumentFileName());
-    }
-
-    @Test
-    public void callToAboutToSubmitHandler_willWriteManuallyUploadedFinalDecisionToCaseForNonDescriptorRoute() throws Exception {
-        setup("callback/writeFinalDecisionManualUploadNonDescriptor.json");
-
-        MockHttpServletResponse response = getResponse(getRequestWithAuthHeader(json, "/ccdAboutToSubmit"));
-        assertHttpStatus(response, HttpStatus.OK);
-        PreSubmitCallbackResponse<SscsCaseData> result = deserialize(response.getContentAsString());
-
-        assertEquals(Collections.EMPTY_SET, result.getErrors());
-
-        assertNull(result.getData().getOutcome());
-
-        assertEquals(DRAFT_DECISION_NOTICE.getValue(), result.getData().getSscsDocument().get(0).getValue().getDocumentType());
-        assertEquals(LocalDate.now().toString(), result.getData().getSscsDocument().get(0).getValue().getDocumentDateAdded());
-        assertEquals("Draft Decision Notice generated on " + LocalDate.now().format(DateTimeFormatter.ofPattern("dd-MM-yyyy")) + ".pdf", result.getData().getSscsDocument().get(0).getValue().getDocumentFileName());
-    }
-}
