+++ src/test/resources/callback/writeFinalDecisionManualUploadNonDescriptor.json
@@ -43,7 +43,7 @@
           }
         },
         "benefitType": {
-          "code": "PIP",
+          "code": "BENEFIT_TYPE",
           "description": null
         },
         "hearingOptions": {
@@ -753,7 +753,7 @@
           }
         },
         "benefitType": {
-          "code": "PIP",
+          "code": "BENEFIT_TYPE",
           "description": null
         },
         "hearingOptions": {
